# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisCreate": {
        "properties": {
          "connection_id": {
            "format": "uuid",
            "title": "Connection Id",
            "type": "string"
          },
          "cv_text": {
            "maxLength": 100000,
            "minLength": 120,
            "title": "Cv Text",
            "type": "string"
          },
          "job_description": {
            "default": "",
            "maxLength": 20000,
            "title": "Job Description",
            "type": "string"
          },
          "target_role": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Target Role",
            "type": "string"
          }
        },
        "required": [
          "cv_text",
          "target_role",
          "connection_id"
        ],
        "title": "AnalysisCreate",
        "type": "object"
      },
      "AnalysisJobView": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "enum": [
              "pending",
              "complete",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "document_id",
          "status",
          "created_at"
        ],
        "title": "AnalysisJobView",
        "type": "object"
      },
      "AnalysisResult": {
        "additionalProperties": false,
        "properties": {
          "career_signals": {
            "items": {
              "$ref": "#/components/schemas/CareerSignal"
            },
            "maxItems": 9,
            "minItems": 9,
            "title": "Career Signals",
            "type": "array"
          },
          "dimensions": {
            "items": {
              "$ref": "#/components/schemas/ReadinessDimension"
            },
            "maxItems": 5,
            "minItems": 5,
            "title": "Dimensions",
            "type": "array"
          },
          "gaps": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "minItems": 1,
            "title": "Gaps",
            "type": "array"
          },
          "lesson_suggestions": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "title": "Lesson Suggestions",
            "type": "array"
          },
          "readiness_score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Readiness Score",
            "type": "integer"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "minItems": 1,
            "title": "Recommendations",
            "type": "array"
          },
          "rewrites": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "title": "Rewrites",
            "type": "array"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "minItems": 1,
            "title": "Strengths",
            "type": "array"
          }
        },
        "required": [
          "readiness_score",
          "dimensions",
          "career_signals",
          "strengths",
          "gaps",
          "recommendations"
        ],
        "title": "AnalysisResult",
        "type": "object"
      },
      "AnswerState": {
        "properties": {
          "answered_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answered At"
          },
          "value": {
            "maximum": 20.0,
            "minimum": 0.0,
            "title": "Value",
            "type": "integer"
          }
        },
        "required": [
          "value"
        ],
        "title": "AnswerState",
        "type": "object"
      },
      "CareerSignal": {
        "properties": {
          "finding": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Finding",
            "type": "string"
          },
          "id": {
            "enum": [
              "decision-velocity",
              "authority-gap",
              "narrative-scarcity",
              "authority-signal",
              "seniority-perception",
              "operational-roi",
              "governance",
              "observability",
              "scalability"
            ],
            "title": "Id",
            "type": "string"
          },
          "label": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "label",
          "score",
          "finding"
        ],
        "title": "CareerSignal",
        "type": "object"
      },
      "CvDocumentCreate": {
        "properties": {
          "consent": {
            "const": true,
            "title": "Consent",
            "type": "boolean"
          },
          "content": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "maxLength": 10485761,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "extracted_text": {
            "anyOf": [
              {
                "maxLength": 100000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extracted Text"
          },
          "filename": {
            "maxLength": 180,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "job_description": {
            "default": "",
            "maxLength": 20000,
            "title": "Job Description",
            "type": "string"
          },
          "mime_type": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Mime Type",
            "type": "string"
          },
          "pasted_text": {
            "anyOf": [
              {
                "maxLength": 100001,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pasted Text"
          },
          "target_role": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Target Role",
            "type": "string"
          }
        },
        "required": [
          "filename",
          "mime_type",
          "target_role",
          "consent"
        ],
        "title": "CvDocumentCreate",
        "type": "object"
      },
      "CvDocumentDetail": {
        "properties": {
          "analyses": {
            "items": {
              "$ref": "#/components/schemas/AnalysisJobView"
            },
            "title": "Analyses",
            "type": "array"
          },
          "document": {
            "$ref": "#/components/schemas/CvDocumentView"
          }
        },
        "required": [
          "document",
          "analyses"
        ],
        "title": "CvDocumentDetail",
        "type": "object"
      },
      "CvDocumentView": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "processing_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Error Code"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "status": {
            "enum": [
              "uploaded",
              "processing",
              "complete",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "mime_type",
          "size_bytes",
          "status",
          "created_at"
        ],
        "title": "CvDocumentView",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LessonProgress": {
        "properties": {
          "answers": {
            "additionalProperties": {
              "$ref": "#/components/schemas/AnswerState"
            },
            "title": "Answers",
            "type": "object"
          },
          "completed": {
            "default": false,
            "title": "Completed",
            "type": "boolean"
          },
          "completion_changed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Changed At"
          },
          "lesson_path": {
            "pattern": "^(phases|certifications)/[a-z0-9][a-z0-9/-]+$",
            "title": "Lesson Path",
            "type": "string"
          },
          "visited_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visited At"
          }
        },
        "required": [
          "lesson_path"
        ],
        "title": "LessonProgress",
        "type": "object"
      },
      "Page": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CvDocumentView"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "offset",
          "limit",
          "total"
        ],
        "title": "Page",
        "type": "object"
      },
      "ProgressState": {
        "properties": {
          "lessons": {
            "items": {
              "$ref": "#/components/schemas/LessonProgress"
            },
            "maxItems": 600,
            "title": "Lessons",
            "type": "array"
          }
        },
        "title": "ProgressState",
        "type": "object"
      },
      "ProviderConnectionView": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "key_hint": {
            "title": "Key Hint",
            "type": "string"
          },
          "model_id": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Model Id",
            "type": "string"
          },
          "provider_id": {
            "$ref": "#/components/schemas/ProviderId"
          }
        },
        "required": [
          "provider_id",
          "model_id",
          "id",
          "key_hint",
          "created_at"
        ],
        "title": "ProviderConnectionView",
        "type": "object"
      },
      "ProviderId": {
        "enum": [
          "gemini",
          "openai",
          "anthropic"
        ],
        "title": "ProviderId",
        "type": "string"
      },
      "ProviderModelUpdate": {
        "properties": {
          "model_id": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Model Id",
            "type": "string"
          }
        },
        "required": [
          "model_id"
        ],
        "title": "ProviderModelUpdate",
        "type": "object"
      },
      "ReadinessDimension": {
        "properties": {
          "evidence": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "title": "Evidence",
            "type": "array"
          },
          "gaps": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "title": "Gaps",
            "type": "array"
          },
          "id": {
            "enum": [
              "role-alignment",
              "evidence",
              "impact",
              "skills",
              "clarity"
            ],
            "title": "Id",
            "type": "string"
          },
          "label": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "rationale": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Rationale",
            "type": "string"
          },
          "score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "label",
          "score",
          "rationale"
        ],
        "title": "ReadinessDimension",
        "type": "object"
      },
      "SaveProviderConnection": {
        "properties": {
          "credential": {
            "format": "password",
            "title": "Credential",
            "type": "string",
            "writeOnly": true
          },
          "model_id": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Model Id",
            "type": "string"
          },
          "provider_id": {
            "$ref": "#/components/schemas/ProviderId"
          }
        },
        "required": [
          "provider_id",
          "model_id",
          "credential"
        ],
        "title": "SaveProviderConnection",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Codeology API experiment",
    "version": "0.1.0-experiment"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/cv/documents": {
      "get": {
        "operationId": "list_documents_api_v1_cv_documents_get",
        "parameters": [
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Documents"
      },
      "post": {
        "operationId": "create_document_api_v1_cv_documents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CvDocumentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CvDocumentView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Document"
      }
    },
    "/api/v1/cv/documents/{document_id}": {
      "delete": {
        "operationId": "delete_document_api_v1_cv_documents__document_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Document"
      },
      "get": {
        "operationId": "get_document_api_v1_cv_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CvDocumentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Document"
      }
    },
    "/api/v1/cv/documents/{document_id}/analyses": {
      "post": {
        "operationId": "analyze_api_v1_cv_documents__document_id__analyses_post",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Document Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisJobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze"
      }
    },
    "/api/v1/health": {
      "get": {
        "operationId": "health_api_v1_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Api V1 Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/progress/reconcile": {
      "post": {
        "operationId": "reconcile_api_v1_progress_reconcile_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressState"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reconcile"
      }
    },
    "/api/v1/providers": {
      "get": {
        "operationId": "providers_api_v1_providers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProviderConnectionView"
                  },
                  "title": "Response Providers Api V1 Providers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Providers"
      },
      "post": {
        "operationId": "save_provider_api_v1_providers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveProviderConnection"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderConnectionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Provider"
      }
    },
    "/api/v1/providers/{connection_id}": {
      "delete": {
        "operationId": "delete_provider_api_v1_providers__connection_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Provider"
      }
    },
    "/api/v1/providers/{connection_id}/model": {
      "patch": {
        "operationId": "update_provider_model_api_v1_providers__connection_id__model_patch",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Connection Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderModelUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderConnectionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Provider Model"
      }
    },
    "/api/v1/readiness": {
      "get": {
        "operationId": "readiness_api_v1_readiness_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Readiness Api V1 Readiness Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Readiness",
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/version": {
      "get": {
        "operationId": "version_api_v1_version_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Version Api V1 Version Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Version",
        "tags": [
          "health"
        ]
      }
    }
  }
}