# apps/web/lib/api/generated.ts
// Generated by apps/api/scripts/generate_contracts.py. Do not edit.
export const OPENAPI_SHA256 = "5e406e6b77172e74cea648aaf5f533cb09426df4ecd9e8b106f77ebbdaa9aa92";
export type ProviderId = "gemini" | "openai" | "anthropic";
export type DocumentStatus = "uploaded" | "processing" | "complete" | "failed";
export type AnalysisStatus = "pending" | "complete" | "failed";
export interface ApiErrorEnvelope { code: string; message: string; request_id: string; }
export interface ProgressAnswer { value: number; answered_at: string | null; }
export interface LessonProgress { lesson_path: string; answers: Record<string, ProgressAnswer>; completed: boolean; completion_changed_at: string | null; visited_at: string | null; }
export interface ProgressState { lessons: LessonProgress[]; }